Store digest of latest image in ImagePolicy's status
The new API field `.status.latestDigest` in the `ImagePolicy` kind
stores the digest of the image referred to by the the
`.status.latestImage` field.

The setting of this field is governed by the newly introduced field
`.spec.digestReflectionPolicy` which takes one of the values `Always`
or `IfNotPresent`. See the updated documentation under `docs/spec/`
for details.

The new status field can be used to pin an image to an immutable
descriptor rather than to a potentially moving tag, increasing the
security of workloads deployed on a cluster.

The goal is to make use of the digest in IAC so that manifests can be
updated with the actual image digest.

Signed-off-by: Max Jonas Werner <mail@makk.es>